1차 수정
$client$
user_action.js
- kakaoLogin() 함수생성, 내용은 미구현

public폴더에 image폴더 추가
- 앞으로 다양한 이미지 넣을 거라는 생각에 단순히 추가 ==> 프론트 맘이므로 지워도 괜찮습니다.

LoginPage.js에 카카오톡 로그인 버튼 뷰 구현
-kakaoLogin()으로 세부 동작 구현할 예정
-image폴더에서 이미지 소스 사용

$server$
index파일에 분류없이 코드가 모여있었기에 라우터를 이용해서 파일과 url 분류

app.js파일 생성
-보통 시작 파일을 app.js로 많이 하는 듯합니다.

index.js파일 분리 (라우터이용, 절대경로: "/" 할당 ==> 기존 유지)
-기존 index파일 내의 함수들을 각 url에 맞게 폴더로 구분하여, 기능별로 파일 세분화

api.js파일 생성 (라우터이용, 절대경로: "/api" 할당 ==> 기존 유지)
-auth,login,logout,register파일들의 라우터 접근을 위한 미들웨어역할 수행

auth,login,logout,register.js파일 생성 (라우터이용, 절대경로: "/api/각파일명" 할당 ==> 기존유지)
-index에 구현되어 있던 각 함수 기능에 따라 파일로 분류해서 구현

kakaologin.js(models폴더 내부)파일 생성
-client에서의 카카오 로그인 요청을 처리할 함수 생성할 목적으로 만듬 ==> 아직 미구현

env.js(config폴더 내부)파일 생성
-kakaologin.js내의 함수에서 쓰일 리다이렉트url, Rest API Key 등등의 설정값을 저장하기 위해 생성

**몇몇 경고 수정
1. 몽고 db 설정 문제 ==> mongoose.set('strictQuery', false); 추가로 해결
