--- Backend(tjdeo)/server/views/error.pug
@@ -1,8 +1,0 @@
-html
-    head
-        e error
-    body 
-        h2 ERROR
-        div.Red = ERROR;
-
-//-error에 대한 페이지 디자인
