--- Backend(tjdeo)/server/views/main.pug
@@ -1,10 +1,0 @@
-html
-    head
-        link(rel="stylesheet" href="/public/index.css")
-    body 
-        h1 User profile page
-        //- div=msg
-        h2 Nickname
-        div.green= JSON.stringify(rows)
-
-//-main에 대한 페이지 디자인
