--- Backend(tjdeo)/server/views/user.pug
@@ -1,10 +1,0 @@
-html
-    head
-        link(rel="stylesheet" href="/public/index.css")
-    body 
-        h1 User Inf
-        //- div=msg
-        h2 Name------Year
-        div.green= JSON.stringify(rows)
-
-//-user에 대한 페이지 디자인
