+++ client/src/_actions/user_action.js
@@ -11,3 +11,7 @@ export function loginUser(dataToSubmit) {
     payload: request,
   };
 }
+
+export function kakaoLogin() {
+  return {};
+} //카카오 버튼 이벤트 발생시(클릭시) 처리하는 함수(단순하게 세부구현 백엔드에게 요청할 예정)
